Update ctr containers and tasks command
This moves container and tasks commands under the containers and tasks
top level commands.
