+++ cmd/ctr/container_delete.go
